Adicionado visualizacoes para os valores das entidades

+++ docker/dbt/models/stg/stg_depreciacao.sql
@@ -19,6 +19,6 @@ with depreciacao as(
 
 select
   *,
-  concat(mes_lancamento,ano_lancamento) as mes_ano_lancamento
+  concat(mes_lancamento,ano_lancamento)::integer as mes_ano_lancamento
 from depreciacao
 where ano_lancamento < 2017
