+++ docker/dbt/models/stg/stg_pensionistas.sql
@@ -17,6 +17,6 @@ with pensionistas as (
 
 select
 	*,
-	concat(mes_lancamento, ano_lancamento) as mes_ano_lancamento
+	concat(mes_lancamento, ano_lancamento)::integer as mes_ano_lancamento
 from pensionistas
 where ano_lancamento < 2017
