+++ docker/dbt/models/stg/stg_pessoal_ativo.sql
@@ -31,6 +31,6 @@ with pessoal_ativo as(
 
 select
 	*,
-	concat(mes_lancamento, ano_lancamento) as mes_ano_lancamento
+	concat(mes_lancamento, ano_lancamento)::integer as mes_ano_lancamento
 from pessoal_ativo
 where ano_lancamento < 2017
